requirements.txt
streamlit
openai
whisper
pytesseract
python-pptx
pandas
PyPDF2
docx2txt
mammoth
Pillow
python-docx
python-magic
